remove LigandAtomMapping and visualization submodule
these are now in gufe
