+++ openfe/setup/atom_mapping/__init__.py
@@ -1,5 +1,5 @@
+from gufe import LigandAtomMapping
 from .ligandatommapper import LigandAtomMapper
-from .ligandatommapping import LigandAtomMapping
 
 from .lomap_mapper import LomapAtomMapper
 from .perses_mapper import PersesAtomMapper
