+++ openfecli/commands/atommapping.py
@@ -74,7 +74,7 @@ def atommapping_print_dict_main(mapper, molA, molB):
 
 def atommapping_visualize_main(mapper, molA, molB, file, ext):
     from rdkit.Chem import Draw
-    from openfe.utils import visualization
+    from gufe.visualization import mapping_visualization as vis
 
     mapping = generate_mapping(mapper, molA, molB)
     ext_to_artist = {
@@ -88,9 +88,9 @@ def atommapping_visualize_main(mapper, molA, molB, file, ext):
             "supported: " + ", ".join([f"'{ext}'" for ext in ext_to_artist])
         )
 
-    contents = visualization.draw_mapping(mapping.componentA_to_componentB,
-                                          mapping.componentA.to_rdkit(),
-                                          mapping.componentB.to_rdkit(), d2d=artist)
+    contents = vis.draw_mapping(mapping.componentA_to_componentB,
+                                mapping.componentA.to_rdkit(),
+                                mapping.componentB.to_rdkit(), d2d=artist)
 
     file.write(contents)
